Add files via upload
Snake Game using HTML, CSS, and JavaScript
This is a simple implementation of the classic Snake game using HTML, CSS, and JavaScript. The game is played in a browser window and the objective is to control a snake, which moves around the game board, eating food and growing in size. The game ends when the snake hits a wall or its own body.

Getting Started
To run the game, you can simply open the index.html file in a web browser. You can also copy the code into your own HTML file or integrate it into an existing website.

Game Rules
The rules of the game are simple:

Use the arrow keys on your keyboard to control the movement of the snake.
The snake will move continuously in the direction you choose.
The objective is to eat the food that appears on the game board.
Eating food will cause the snake to grow in length.
If the snake hits a wall or its own body, the game ends.
Game Components
The game is made up of the following components:

The game board, which is a rectangular grid.
The snake, which is a series of connected blocks.
The food, which appears randomly on the game board.
The score, which keeps track of the player's score.
Implementation Details
The game is implemented using HTML, CSS, and JavaScript. The game board and snake are created using HTML div elements, which are styled using CSS. The movement of the snake is controlled using JavaScript, which updates the position of the snake blocks on the game board. The food is randomly generated using JavaScript, and the score is updated using JavaScript as well.

Credits
This implementation of the Snake game was created by [Vinay Kumar]. It is based on the classic Snake game, which was first released in the 1970s.